Refactor mquery to use ES6 module syntax
The included mquery library uses a non-standard export syntax. This makes webpack unable to transpile it in the angular2 build process. This PR updates the mquery exports.

With this fixed, angular2 will use the RxDB source files, transpile them in it's own build process, and will not require the babel-polyfill.