Fix react-native example project
The default plugin for validation, `is-my-json-valid`, depends on
`generate-functions`, which uses core Node libraries and is therefore not usable
outside of the core environment. Attempting the build the project before led
to the following error:

    Unable to resolve module `util` from `<workspace/node_modules/generate_function/index.js`: Module `util` does not exist in the Haste module map

The solution to this is to use AJV for validation instead.

Also added try/catch around await statements, since that was the only way I
could find this problem in the first place.