Update expo version in app.json
Without this, we get a "React Native Version Mismatch" error. I think the issue
is that @expo/xdl has a mapping from SDK versions to RN versions, and specifying
the wrong SDK here ends up indicating the wrong JS RN version later. It's
strange to me that the version is not taken from package.json, though. Why is
this field even here?