Bug: compound index strings calculate wrong bounds
There seems to be a bug in some RxDB storages (IndexedDB, Memory) when a compound index covers more than all fields of a selector (one or more additional fields are present in comparison to the fields used in the query selector). This can lead to situations where the query planner sets the queryPlan.selectorSatisfiedByIndex to true, but the bounds (upper/lower) might be wrong (in our case a boolean integer was set to a range between 0-9). When the additional index field is more significant than the actual fields used in the selector, the upper bound is too high, and the number of documents returned is greater than the expected number.

The storage implementations assume that queryPlan.selectorSatisfiedByIndex=true implies that the range defined by the bounds includes all valid documents, and skips the queryMatcher logic.